+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 * nostr: don't set root tags when the root is null ([Yuki Kishimoto])
 * nostr: update `RelayMessage::NegErr` variant ([Yuki Kishimoto])
 * nostr: accept either `EventBuilder` or `UnsignedEvent` as rumor in NIP59 functions ([Yuki Kishimoto] and [arkanoider])
+* nostr: require `fmt::Debug`, `Send` and `Sync` for `NostrSigner` ([Yuki Kishimoto])
 * database: add manual trait implementations for `BTreeCappedSet` ([Yuki Kishimoto]) 
 * database: replace LRU with custom memory cache for IDs tracking ([Yuki Kishimoto])
 * lmdb: use `async-utility` to spawn blocking tasks ([Yuki Kishimoto])